distutils-r1.eclass: Minimize & unify dift --via-venv logic
Modify `distutils_install_for_testing --via-venv` to create a minimal
venv manually rather than relying on Python to do so.  Use root-style
install rather than the egg-style to improve consistency with regular
installs.

This is a step towards unifying different install layouts used within
the eclass.  Right now we support three different variants for testing:

1. The build-dir layout that's created by python_compile() and exposed
   unconditionally through PYTHONPATH.

2. The --via-root layout of dift that resembles install closer
   (primarily through including package metadata) and also uses
   PYTHONPATH.

3. The --via-venv layout of dift that creates a venv and installs
   the packages there.  It requires only PATH, not PYTHONPATH.

The last layout is the newest and probably the most compatible but it
requires additional install step.  Since the PEP517 build logic is going
to require installing a wheel anyway, the plan is to inject a minimal
venv into the staging directory and use it unconditionally for tests.

The purpose of this patch is to prepare a single code snippet that will
be used both by dift and the new logic logic.

Signed-off-by: Michał Górny <mgorny@gentoo.org>