python-utils-r1.eclass: epytest, pass --force-short-summary
Pass `--force-short-summary` when pytest is new enough to accept it,
in order to shorten the long (possibly multi-line) exception messages
that make the "short summary" hard to read.  The complete exception
messages are available along with the backtraces anyway.

This option is available with >=dev-python/pytest-8.4.0, and it restores
the behavior of older dev-python/pytest versions.

Signed-off-by: Michał Górny <mgorny@gentoo.org>